requirements.txt
# Core dependencies
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
ultralytics>=8.0.0

# Optional (for enhanced features)
scikit-image>=0.19.0  # For image processing
matplotlib>=3.5.0     # For plotting statistics
pandas>=1.4.0         # For Excel export
reportlab>=3.6.0      # For PDF export (optional)